refactor/move get_anns to pecha
